# apps/web/src/lib/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RouteHub API",
    "version": "1.0.0",
    "description": "API documentation for RouteHub — fleet management, payments, documents, and inspections."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://geosafety.routehub.ge",
      "description": "Production (master)"
    },
    {
      "url": "https://team2.routehub.ge",
      "description": "Team 2"
    },
    {
      "url": "https://team3.routehub.ge",
      "description": "Team 3"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie (set automatically after login)"
      },
      "cronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Bearer token matching CRON_SECRET env var (for scheduled jobs)"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/documents/templates": {
      "get": {
        "summary": "List document templates for a board",
        "tags": ["Documents"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "boardId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Board ID to fetch templates for (also returns global templates with null board_id)"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of active document templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing boardId parameter"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Upload a new document template",
        "tags": ["Documents"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file", "name"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Template file (.docx, .xlsx, or .xls)"
                  },
                  "name": {
                    "type": "string",
                    "description": "Display name for the template"
                  },
                  "description": {
                    "type": "string",
                    "description": "Optional template description"
                  },
                  "boardId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Board to associate template with (null for global)"
                  },
                  "workspaceId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Workspace ID for storage path"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template created successfully (may include warning about tag parsing)"
          },
          "400": {
            "description": "Missing file/name or unsupported file type"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/documents/templates/{id}": {
      "get": {
        "summary": "Get a single document template by ID",
        "tags": ["Documents"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Template ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Template object"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update a document template",
        "tags": ["Documents"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Template ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "tag_mapping": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Mapping of template tags to column IDs or computed fields"
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated template object"
          },
          "400": {
            "description": "No valid fields to update"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a document template and its storage file",
        "tags": ["Documents"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Template ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Template deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}